backend-ai/app/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-chroma
openai
chromadb
pydantic
requests
python-dotenv
numpy
scikit-learn
